+++ tests/Fixtures/SearchablePost.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace LaravelScoutDebouncer\Tests\Fixtures;
+namespace JonHassall\ScoutBatcher\Tests\Fixtures;
 
 use Illuminate\Database\Eloquent\Model;
-use LaravelScoutDebouncer\Concerns\BatchSearchable;
+use JonHassall\ScoutBatcher\Concerns\BatchSearchable;
 
 class SearchablePost extends Model
 {